style: :art: Style solidity code
Style solidity code

+++ contracts/TokenVesting.sol
@@ -1,10 +1,8 @@
 pragma solidity ^0.8.0;
 
-
 import "./interfaces/IStakeManager.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts/security/ReentrancyGuard.sol";
-
 
 /**
  * @title TokenVesting
@@ -25,7 +23,7 @@ contract TokenVesting is ReentrancyGuard {
     event TokensReleased(IERC20 indexed token, uint256 amount);
     event TokenVestingRevoked(IERC20 indexed token);
 
-    uint public constant CLIFF_DENOMINATOR = 5; // x / 5 = 20% of x
+    uint256 public constant CLIFF_DENOMINATOR = 5; // x / 5 = 20% of x
 
     // beneficiary of tokens after they are released
     address public immutable beneficiary;
@@ -35,17 +33,17 @@ contract TokenVesting is ReentrancyGuard {
     bool public immutable revocable;
 
     // Durations and timestamps are expressed in UNIX time, the same units as block.timestamp.
-    uint public immutable start;
-    uint public immutable cliff;
-    uint public immutable end;
+    uint256 public immutable start;
+    uint256 public immutable cliff;
+    uint256 public immutable end;
 
     // If false, staking is not allowed
     bool public immutable canStake;
     // The staking contract to stake to if `canStake`
     IStakeManager public immutable stakeManager;
 
-    mapping (IERC20 => uint256) public released;
-    mapping (IERC20 => bool) public revoked;
+    mapping(IERC20 => uint256) public released;
+    mapping(IERC20 => bool) public revoked;
 
     /**
      * @dev Creates a vesting contract that vests its balance of any ERC20 token to the
@@ -66,13 +64,13 @@ contract TokenVesting is ReentrancyGuard {
      * @param canStake_ whether the investor is allowed to use vested funds to stake
      * @param stakeManager_ the staking contract to stake to if canStake
      */
-    constructor (
+    constructor(
         address beneficiary_,
         address revoker_,
         bool revocable_,
-        uint start_,
-        uint cliff_,
-        uint end_,
+        uint256 start_,
+        uint256 cliff_,
+        uint256 end_,
         bool canStake_,
         IStakeManager stakeManager_
     ) {
@@ -103,7 +101,7 @@ contract TokenVesting is ReentrancyGuard {
      * @param nodeID the nodeID to stake for.
      * @param amount the amount to stake out of the current funds in this contract.
      */
-    function stake(bytes32 nodeID, uint amount) external {
+    function stake(bytes32 nodeID, uint256 amount) external {
         require(msg.sender == beneficiary, "TokenVesting: not the beneficiary");
         require(canStake, "TokenVesting: cannot stake");
 
@@ -126,7 +124,6 @@ contract TokenVesting is ReentrancyGuard {
 
         emit TokensReleased(token, unreleased);
     }
-    
 
     /**
      * @notice Allows the revoker to revoke the vesting.
@@ -140,12 +137,12 @@ contract TokenVesting is ReentrancyGuard {
         require(msg.sender == revoker, "TokenVesting: not the revoker");
         require(revocable, "TokenVesting: cannot revoke");
         require(!revoked[token], "TokenVesting: token already revoked");
-        require(block.timestamp <= end , "TokenVesting: vesting period expired");
+        require(block.timestamp <= end, "TokenVesting: vesting period expired");
 
-        uint balance = token.balanceOf(address(this));
+        uint256 balance = token.balanceOf(address(this));
 
-        uint unreleased = _releasableAmount(token);
-        uint refund = balance - unreleased;
+        uint256 unreleased = _releasableAmount(token);
+        uint256 refund = balance - unreleased;
 
         revoked[token] = true;
 
@@ -158,18 +155,18 @@ contract TokenVesting is ReentrancyGuard {
      * @notice Allows the revoker to retrieve tokens that have been unstaked
      *         after the revoke function has been called (in canStake contracts).
      *         Safeguard mechanism in case of unstaking happening after revoke.
-     *         Otherwise funds would be locked. In !canStake contracts all the 
+     *         Otherwise funds would be locked. In !canStake contracts all the
      *         funds are withdrawn once revoked is called, so no need for this
      * @param token ERC20 token which is being vested.
      */
-    function retrieveRevokedFunds (IERC20 token) external {
+    function retrieveRevokedFunds(IERC20 token) external {
         require(msg.sender == revoker, "TokenVesting: not the revoker");
         require(revocable, "TokenVesting: cannot revoke");
-        require(revoked[token], "TokenVesting: token not revoked");  
+        require(revoked[token], "TokenVesting: token not revoked");
 
         require(canStake, "TokenVesting: not retrievable");
 
-        uint balance = token.balanceOf(address(this));
+        uint256 balance = token.balanceOf(address(this));
 
         token.safeTransfer(revoker, balance);
     }
@@ -178,7 +175,7 @@ contract TokenVesting is ReentrancyGuard {
      * @dev Calculates the amount that has already vested but hasn't been released yet.
      * @param token ERC20 token which is being vested.
      */
-    function _releasableAmount(IERC20 token) private view returns (uint) {
+    function _releasableAmount(IERC20 token) private view returns (uint256) {
         return _vestedAmount(token) - released[token];
     }
 
@@ -187,8 +184,7 @@ contract TokenVesting is ReentrancyGuard {
      *      option A, full unvesting at the end for contract B.
      * @param token ERC20 token which is being vested.
      */
-    function _vestedAmount(IERC20 token) private view returns (uint) {
-        
+    function _vestedAmount(IERC20 token) private view returns (uint256) {
         uint256 currentBalance = token.balanceOf(address(this));
         uint256 totalBalance = currentBalance + released[token];
 
@@ -198,7 +194,7 @@ contract TokenVesting is ReentrancyGuard {
             return totalBalance;
         } else {
             // should never enter this if canStake == true, since cliff == end
-            assert (!canStake);
+            assert(!canStake);
             uint256 cliffAmount = totalBalance / CLIFF_DENOMINATOR;
             return cliffAmount + (totalBalance - cliffAmount) * (block.timestamp - cliff) / (end - cliff);
         }
