+++ contracts/interfaces/IShared.sol
@@ -1,33 +1,31 @@
 pragma solidity ^0.8.0;
 
-
 /**
-* @title    Shared interface
-* @notice   Holds structs needed by other interfaces
-* @author   Quantaf1re (James Key)
-*/
+ * @title    Shared interface
+ * @notice   Holds structs needed by other interfaces
+ * @author   Quantaf1re (James Key)
+ */
 interface IShared {
-
     /**
-    * @dev  SchnorrSECP256K1 requires that each key has a public key part (x coordinate),
-    *       a parity for the y coordinate (0 if the y ordinate of the public key is even, 1
-    *       if it's odd)
-    */
+     * @dev  SchnorrSECP256K1 requires that each key has a public key part (x coordinate),
+     *       a parity for the y coordinate (0 if the y ordinate of the public key is even, 1
+     *       if it's odd)
+     */
     struct Key {
-        uint pubKeyX;
+        uint256 pubKeyX;
         uint8 pubKeyYParity;
     }
 
     /**
-    * @dev  Contains a signature and the msgHash that the signature is over. Kept as a single
-    *       struct since they should always be used together
-    */
+     * @dev  Contains a signature and the msgHash that the signature is over. Kept as a single
+     *       struct since they should always be used together
+     */
     struct SigData {
         address keyManAddr;
-        uint chainID;
-        uint msgHash;
-        uint sig;
-        uint nonce;
+        uint256 chainID;
+        uint256 msgHash;
+        uint256 sig;
+        uint256 nonce;
         address kTimesGAddr;
     }
 }
